Move `Header` and constituents to `ergo-chain-types`
