+++ ergo-chain-types/Cargo.toml
@@ -26,6 +26,8 @@ serde_json = { version = "1.0", optional = true }
 url = "2.2"
 k256 = { version = "0.10.2", features = ["arithmetic", "ecdsa"] }
 elliptic-curve = {version = "0.11.6", features = [ "ff"]}
+num-bigint = "0.4.0"
+byteorder = "1"
 
 [dependencies.serde_with]
 version = "1.9.1"
