+++ ergo-chain-types/src/json.rs
@@ -1,0 +1,4 @@
+//! JSON serialization
+
+pub(crate) mod autolykos_solution;
+pub(crate) mod votes;
