+++ ergo-chain-types/src/json/autolykos_solution.rs
