+++ ergo-chain-types/src/lib.rs
@@ -22,8 +22,11 @@ mod block_id;
 mod digest32;
 pub mod ec_point;
 mod extensioncandidate;
+mod header;
+mod json;
 mod peer_addr;
 mod peer_connection_dir;
+mod votes;
 
 pub use base16_bytes::Base16DecodedBytes;
 pub use base16_bytes::Base16EncodedBytes;
@@ -33,6 +36,9 @@ pub use digest32::ADDigest;
 pub use digest32::Digest;
 pub use digest32::Digest32;
 pub use digest32::DigestNError;
+pub use ec_point::EcPoint;
 pub use extensioncandidate::ExtensionCandidate;
+pub use header::{AutolykosSolution, Header};
 pub use peer_addr::PeerAddr;
 pub use peer_connection_dir::ConnectionDirection;
+pub use votes::Votes;
