+++ ergotree-ir/src/chain/preheader.rs
@@ -1,10 +1,6 @@
 //! Block header with fields that can be predicted by miner
-
-use ergo_chain_types::BlockId;
-
 use crate::sigma_protocol::dlog_group;
-
-use super::votes::Votes;
+use ergo_chain_types::{BlockId, Header, Votes};
 
 /// Block header with the current `spendingTransaction`, that can be predicted
 /// by a miner before it's formation
@@ -24,6 +20,20 @@ pub struct PreHeader {
     pub miner_pk: Box<dlog_group::EcPoint>,
     /// Votes
     pub votes: Votes,
+}
+
+impl From<Header> for PreHeader {
+    fn from(bh: Header) -> Self {
+        PreHeader {
+            version: bh.version,
+            parent_id: bh.parent_id,
+            timestamp: bh.timestamp,
+            n_bits: bh.n_bits,
+            height: bh.height,
+            miner_pk: bh.autolykos_solution.miner_pk,
+            votes: bh.votes,
+        }
+    }
 }
 
 #[cfg(feature = "arbitrary")]
